pi/requirements.txt
Adafruit-Blinka==8.69.0
adafruit-circuitpython-bme280==2.6.30
adafruit-circuitpython-busdevice==5.2.15
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-typing==1.12.3
Adafruit-PlatformDetect==3.86.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
board==1.0
iniconfig==2.3.0
packaging==26.0
paho-mqtt==2.1.0
pluggy==1.6.0
pyftdi==0.57.1
Pygments==2.19.2
pyserial==3.5
pytest==9.0.2
python-dotenv==1.2.1
pyusb==1.3.1
sysv_ipc==1.2.0
typing_extensions==4.15.0
